fix: 🐛 formal analyzer context intefaces and processing touch-ups (#6)
fix: 🐛 formal analyzer context interfaces and processing touch-ups

This commit overhauls the interface between analyzer modules to remove the dependency on in-memory data, allowing for more flexible input/output handling.

- Previously, analyzer functions operated with Polars DataFrames as input/output, which constrained analysis to in-memory operations.
- This refactor introduces analysis contexts, enabling modules to load, scan, stream files, or write output in chunks, as well as use temporary directories as needed.
- The new design allows analyzer-specific handling of larger-than-memory datasets.
- Parquet remains the standard data exchange format between analyzers.
- Current utility functions still rely on Polars DataFrame/LazyFrame, but the system is capable of supporting Pandas if needed. However, this has not been implemented as all analyzers currently use Polars.
- Secondary analyzers and web presenters can now depend on other secondary analyzers. This allows us to de-duplicate logic where the same kind of calculation is needed to produce tables that are both exported and shown in the web dashboard.
- The refactor introduces some verbosity due to added IO logic, but this is necessary for future optimization and handling of large datasets across analyzers.

This change sets up the foundation for more scalable analysis workflows moving forward.

The above also allows for a snappier data import flow because we no longer need to load in the
data fully. Specifically, CSV import now uses CSV scanning, which is easier on the RAM.

A couple of cosmetic touch-ups of the table renderer are included.